backend/layers/shared/python/requirements.txt
boto3==1.35.0
botocore==1.35.0
PyJWT==2.10.1
requests==2.32.4
numpy==1.26.4
scipy==1.11.4
setuptools>=70.0.0
